Edit_video_CLI/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edit_video_cli"
version = "0.1.0"
authors = [
    {name = "Diego Fornalha", email = "youremail@example.com"},
]
description = "CLI para edição de vídeo com transcrição e geração de SEO"
readme = "README_.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.1",
    "numpy>=1.24.3",
    "moviepy>=1.0.3",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
edit-video = "edit_video_cli.src.main:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/edit_video_cli"
"Bug Tracker" = "https://github.com/yourusername/edit_video_cli/issues" 